backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-dotenv>=1.0.1
openai>=1.51.0
pydantic>=2.10.0
pydantic-settings>=2.5.2
python-jose[cryptography]>=3.3.0
httpx>=0.27.2
sse-starlette>=2.1.3
python-multipart>=0.0.9
pypdf>=4.0.0
langsmith>=0.1.0
